move Logout and remove levenary package
